+++ src/components/_Settings_UserSettings/style.scss
@@ -1,0 +1,8 @@
+.user-settings-modal_small {
+    font-size: 16px;
+    font-weight: bold;
+    margin-bottom: 10px;
+    display: inline-block;
+    color: #a8b1c4;
+    font-style: italic;
+}
