+++ src/components/_User_ChangePassword/index.jsx
@@ -6,15 +6,14 @@ import AuthContext from '../../store/AuthContext'
 import { CheckCircle } from 'react-feather'
 
 export default function ChangePassword(props) {
-
     const contextValue = useContext(AuthContext)
-    const [isLoading, setIsLoading] = useState(true);
-    const [pin, setPin] = useState("");
-    const [error, setError] = useState("");
-    const [resetStep, setResetStep] = useState(1);
-    const [resetToken, setResetToken] = useState("");
-    const [password, setPassword] = useState("");
-    const [confirmPassword, setConfirmPassword] = useState("");
+    const [isLoading, setIsLoading] = useState(true)
+    const [pin, setPin] = useState('')
+    const [error, setError] = useState('')
+    const [resetStep, setResetStep] = useState(1)
+    const [resetToken, setResetToken] = useState('')
+    const [password, setPassword] = useState('')
+    const [confirmPassword, setConfirmPassword] = useState('')
 
     useEffect(() => {
         if (props.isOpenState) {
@@ -30,27 +29,24 @@ export default function ChangePassword(props) {
             fetch(process.env.REACT_APP_API_URL + '/api/auth/reset', requestOptions)
                 .then((response) => response.json())
                 .then((data) => {
-                    console.log('D', data);
                     if (data.message == 'Reset success') {
-                        setIsLoading(false);
-                        setError("");
-                        setResetStep(1);
-                        setResetToken(data.token);
+                        setIsLoading(false)
+                        setError('')
+                        setResetStep(1)
+                        setResetToken(data.token)
                     } else {
-                        setIsLoading(false);
-                        setError("An error occured. Please try again later.");
-                        setResetStep(0);
+                        setIsLoading(false)
+                        setError('An error occured. Please try again later.')
+                        setResetStep(0)
                     }
                 })
         }
     }, [props.isOpenState])
 
     const resetPassword = () => {
-
-        setIsLoading(true);
+        setIsLoading(true)
 
         if (resetStep == 1) {
-
             // Check if pin is correct
             const requestOptions = {
                 method: 'POST',
@@ -65,34 +61,29 @@ export default function ChangePassword(props) {
             fetch(process.env.REACT_APP_API_URL + '/api/auth/reset/2', requestOptions)
                 .then((response) => response.json())
                 .then((data) => {
-                    console.log(data);
                     if (data.message == 'PIN correct') {
-                        setIsLoading(false);
-                        setError("");
-                        setResetStep(2);
+                        setIsLoading(false)
+                        setError('')
+                        setResetStep(2)
                     } else if (data.message == 'PIN incorrect') {
-                        setIsLoading(false);
-                        setError("The PIN you entered is incorrect. Please try again.");
-                        setResetStep(1);
+                        setIsLoading(false)
+                        setError('The PIN you entered is incorrect. Please try again.')
+                        setResetStep(1)
                     }
                 })
-
         } else if (resetStep == 2) {
-
             // Check if passwords match
             if (password != confirmPassword) {
-                setIsLoading(false);
-                setError("The passwords you entered do not match.");
-                setResetStep(2);
+                setIsLoading(false)
+                setError('The passwords you entered do not match.')
+                setResetStep(2)
             } else {
-
                 // Check if password is longer than 8 characters
                 if (password.length < 8) {
-                    setIsLoading(false);
-                    setError("The password you entered is too short. Please enter a password with at least 8 characters.");
-                    setResetStep(2);
+                    setIsLoading(false)
+                    setError('The password you entered is too short. Please enter a password with at least 8 characters.')
+                    setResetStep(2)
                 } else {
-
                     // Reset Password
                     const requestOptions = {
                         method: 'POST',
@@ -109,30 +100,25 @@ export default function ChangePassword(props) {
                     fetch(process.env.REACT_APP_API_URL + '/api/auth/reset/3', requestOptions)
                         .then((response) => response.json())
                         .then((data) => {
-                            console.log(data);
                             if (data.message == 'Password resetted') {
-                                setIsLoading(false);
-                                setError("");
-                                setResetStep(3);
+                                setIsLoading(false)
+                                setError('')
+                                setResetStep(3)
                             } else if (data.message == 'PIN incorrect') {
-                                setIsLoading(false);
-                                setError("The PIN you entered is incorrect. Please try again.");
-                                setResetStep(1);
+                                setIsLoading(false)
+                                setError('The PIN you entered is incorrect. Please try again.')
+                                setResetStep(1)
                             } else {
-                                setIsLoading(false);
-                                setError("An error occured. Please try again later.");
-                                setResetStep(0);
+                                setIsLoading(false)
+                                setError('An error occured. Please try again later.')
+                                setResetStep(0)
                             }
                         })
-
                 }
-
             }
-
         } else {
-            props.isOpenStateController(false);
+            props.isOpenStateController(false)
         }
-
     }
 
     return (
@@ -143,12 +129,11 @@ export default function ChangePassword(props) {
             }}
             open={props.isOpenState}
             size="tiny"
-            className='password-reset-modal'
+            className="password-reset-modal"
         >
             <Modal.Header>Reset your Password</Modal.Header>
             <Modal.Content>
-
-                {error != "" && (
+                {error != '' && (
                     <Message negative>
                         <Message.Header>Warning:</Message.Header>
                         <p>{error}</p>
@@ -157,31 +142,60 @@ export default function ChangePassword(props) {
 
                 {isLoading ? (
                     <div className="password-reset-modal-loader">
-                        <Loader size="medium" active>Loading...</Loader>
+                        <Loader size="medium" active>
+                            Loading...
+                        </Loader>
                     </div>
                 ) : (
                     <>
                         {resetStep === 1 && (
                             <>
-                                <p>
-                                    Please type in the PIN you just received to reset your password.
-                                </p>
+                                <p>Please type in the PIN you just received to reset your password.</p>
                                 <Form.Field>
-                                    <Input autoFocus fluid onChange={(e) => {setPin(e.target.value)}} type="text" value={pin} maxLength="6" placeholder="PIN" className="userPin" />
+                                    <Input
+                                        autoFocus
+                                        fluid
+                                        onChange={(e) => {
+                                            setPin(e.target.value)
+                                        }}
+                                        type="text"
+                                        value={pin}
+                                        maxLength="6"
+                                        placeholder="PIN"
+                                        className="userPin"
+                                    />
                                 </Form.Field>
                             </>
                         )}
 
                         {resetStep === 2 && (
                             <>
-                                <p>
-                                    Please type in your new password below.
-                                </p>
+                                <p>Please type in your new password below.</p>
                                 <Form.Field>
-                                    <Input autoFocus fluid onChange={(e) => {setPassword(e.target.value)}} type="password" minLength="8" value={password} placeholder="New Password" />
-                                </Form.Field><br />
+                                    <Input
+                                        autoFocus
+                                        fluid
+                                        onChange={(e) => {
+                                            setPassword(e.target.value)
+                                        }}
+                                        type="password"
+                                        minLength="8"
+                                        value={password}
+                                        placeholder="New Password"
+                                    />
+                                </Form.Field>
+                                <br />
                                 <Form.Field>
-                                    <Input fluid onChange={(e) => {setConfirmPassword(e.target.value)}} type="password" minLength="8" value={confirmPassword} placeholder="Confirm Password" />
+                                    <Input
+                                        fluid
+                                        onChange={(e) => {
+                                            setConfirmPassword(e.target.value)
+                                        }}
+                                        type="password"
+                                        minLength="8"
+                                        value={confirmPassword}
+                                        placeholder="Confirm Password"
+                                    />
                                 </Form.Field>
                             </>
                         )}
@@ -195,8 +209,6 @@ export default function ChangePassword(props) {
                         )}
                     </>
                 )}
-
-
             </Modal.Content>
             <Modal.Actions>
                 {resetStep !== 3 && (
@@ -211,11 +223,10 @@ export default function ChangePassword(props) {
                 )}
             </Modal.Actions>
         </Modal>
-    );
-
+    )
 }
 
 ChangePassword.propTypes = {
     isOpenState: PropTypes.bool.isRequired,
-    isOpenStateController: PropTypes.func.isRequired
+    isOpenStateController: PropTypes.func.isRequired,
 }
