+++ src/components/_User_ChangePassword/style.scss
@@ -3,7 +3,7 @@
         text-align: center !important;
         letter-spacing: 1em;
         padding-right: 0 !important;
-        font-family: "JetBrains Mono", monospace !important;
+        font-family: 'JetBrains Mono', monospace !important;
     }
 
     &-loader {
